backend/app/utils.py
import os
import requests
import logging
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

logger = logging.getLogger(__name__)
logging.basicConfig(level=logging.INFO)

# --- Placeholder for Financial Data API ---
# In a real application, you would replace this with actual API calls
# to a financial data provider like Alpha Vantage, Financial Modeling Prep, etc.
# You would need to sign up for their API key.

# Example: Alpha Vantage (replace with your actual API key)
# ALPHA_VANTAGE_API_KEY = os.getenv("ALPHA_VANTAGE_API_KEY")
# ALPHA_VANTAGE_BASE_URL = "https://www.alphavantage.co/query"

def get_stock_price(symbol):
    """
    Retrieves the current stock price for a given symbol.
    This is a MOCK function.
    """
    logger.info("Attempting to get stock price for symbol: %s", symbol)
    # Mock data for demonstration
    mock_prices = {
        "AAPL": 170.50,
        "GOOGL": 1800.25,
        "MSFT": 420.10,
        "AMZN": 185.70
    }
    price = mock_prices.get(symbol.upper())
    if price:
        logger.info("Mock stock price for %s: %s", symbol, price)
        return {"symbol": symbol.upper(), "price": price, "currency": "USD"}
    else:
        logger.warning("Mock stock price not found for symbol: %s", symbol)
        return {"error": f"Stock price for {symbol.upper()} not found (mock data)."}

    # # Example of how you might call a real API (e.g., Alpha Vantage)
    # if not ALPHA_VANTAGE_API_KEY:
    #     logger.error("ALPHA_VANTAGE_API_KEY not set in .env")
    #     return {"error": "Financial data API key not configured."}
    # try:
    #     params = {
    #         "function": "GLOBAL_QUOTE",
    #         "symbol": symbol,
    #         "apikey": ALPHA_VANTAGE_API_KEY
    #     }
    #     response = requests.get(ALPHA_VANTAGE_BASE_URL, params=params)
    #     response.raise_for_status() # Raise an exception for HTTP errors
    #     data = response.json()
    #     if "Global Quote" in data:
    #         quote = data["Global Quote"]
    #         price = float(quote["05. price"])
    #         return {"symbol": symbol.upper(), "price": price, "currency": "USD"}
    #     elif "Error Message" in data:
    #         return {"error": data["Error Message"]}
    #     else:
    #         return {"error": "Could not retrieve stock price for " + symbol}
    # except requests.exceptions.RequestException as e:
    #     logger.exception("Error fetching stock price from Alpha Vantage for %s:", symbol)
    #     return {"error": f"Failed to connect to financial data service: {e}"}


def get_currency_exchange_rate(from_currency, to_currency):
    """
    Retrieves the current exchange rate between two currencies.
    This is a MOCK function.
    """
    logger.info("Attempting to get exchange rate from %s to %s", from_currency, to_currency)
    # Mock data for demonstration
    mock_rates = {
        "USD_LKR": 300.00,
        "EUR_USD": 1.08,
        "GBP_USD": 1.27,
        "JPY_USD": 0.0064,
        "LKR_USD": 0.0033
    }
    rate_key = f"{from_currency.upper()}_{to_currency.upper()}"
    rate = mock_rates.get(rate_key)
    if rate:
        logger.info("Mock exchange rate for %s: %s", rate_key, rate)
        return {"from": from_currency.upper(), "to": to_currency.upper(), "rate": rate}
    else:
        logger.warning("Mock exchange rate not found for %s", rate_key)
        return {"error": f"Exchange rate for {from_currency.upper()} to {to_currency.upper()} not found (mock data)."}

    # # Example of how you might call a real API (e.g., Alpha Vantage)
    # if not ALPHA_VANTAGE_API_KEY:
    #     logger.error("ALPHA_VANTAGE_API_KEY not set in .env")
    #     return {"error": "Financial data API key not configured."}
    # try:
    #     params = {
    #         "function": "CURRENCY_EXCHANGE_RATE",
    #         "from_currency": from_currency,
    #         "to_currency": to_currency,
    #         "apikey": ALPHA_VANTAGE_API_KEY
    #     }
    #     response = requests.get(ALPHA_VANTAGE_BASE_URL, params=params)
    #     response.raise_for_status()
    #     data = response.json()
    #     if "Realtime Currency Exchange Rate" in data:
    #         rate_info = data["Realtime Currency Exchange Rate"]
    #         rate = float(rate_info["5. Exchange Rate"])
    #         return {"from": from_currency.upper(), "to": to_currency.upper(), "rate": rate}
    #     elif "Error Message" in data:
    #         return {"error": data["Error Message"]}
    #     else:
    #         return {"error": "Could not retrieve exchange rate for " + from_currency + " to " + to_currency}
    # except requests.exceptions.RequestException as e:
    #     logger.exception("Error fetching exchange rate from Alpha Vantage for %s to %s:", from_currency, to_currency)
    #     return {"error": f"Failed to connect to financial data service: {e}"}

