backend/requirements.txt
Flask
Flask-CORS
pyodbc
python-dotenv
pytest
openai
